+++ talk/media/webrtc/fakewebrtcvideoengine.h
@@ -175,7 +175,7 @@ class FakeWebRtcVideoEncoder : public webrtc::VideoEncoder {
   }
 
   virtual int32 SetChannelParameters(uint32 packetLoss,
-                                     int rtt) {
+                                     int64_t rtt) {
     return WEBRTC_VIDEO_CODEC_OK;
   }
 
@@ -1134,15 +1134,15 @@ class FakeWebRtcVideoEngine
   WEBRTC_STUB_CONST(GetRtcpPacketTypeCounters, (int,
       webrtc::RtcpPacketTypeCounter*, webrtc::RtcpPacketTypeCounter*));
   WEBRTC_STUB_CONST(GetReceivedRTCPStatistics, (const int, unsigned short&,
-      unsigned int&, unsigned int&, unsigned int&, int&));
+      unsigned int&, unsigned int&, unsigned int&, int64_t&));
   WEBRTC_STUB_CONST(GetSentRTCPStatistics, (const int, unsigned short&,
-      unsigned int&, unsigned int&, unsigned int&, int&));
+      unsigned int&, unsigned int&, unsigned int&, int64_t&));
   WEBRTC_STUB_CONST(GetRTPStatistics, (const int, size_t&, unsigned int&,
       size_t&, unsigned int&));
   WEBRTC_STUB_CONST(GetReceiveChannelRtcpStatistics, (const int,
-      webrtc::RtcpStatistics&, int&));
+      webrtc::RtcpStatistics&, int64_t&));
   WEBRTC_STUB_CONST(GetSendChannelRtcpStatistics, (const int,
-      webrtc::RtcpStatistics&, int&));
+      webrtc::RtcpStatistics&, int64_t&));
   WEBRTC_STUB_CONST(GetRtpStatistics, (const int, webrtc::StreamDataCounters&,
       webrtc::StreamDataCounters&));
   WEBRTC_FUNC_CONST(GetBandwidthUsage, (const int channel,
