+++ webrtc/modules/interface/module_common_types.h
@@ -279,7 +279,7 @@ struct FecProtectionParams {
 // CallStats object using RegisterStatsObserver.
 class CallStatsObserver {
  public:
-  virtual void OnRttUpdate(uint32_t rtt_ms) = 0;
+  virtual void OnRttUpdate(int64_t rtt_ms) = 0;
 
   virtual ~CallStatsObserver() {}
 };
