+++ webrtc/modules/remote_bitrate_estimator/aimd_rate_control.h
@@ -39,7 +39,7 @@ class AimdRateControl : public RemoteRateControl {
       int64_t time_now, uint32_t incoming_bitrate_bps) const OVERRIDE;
   virtual uint32_t LatestEstimate() const OVERRIDE;
   virtual uint32_t UpdateBandwidthEstimate(int64_t now_ms) OVERRIDE;
-  virtual void SetRtt(uint32_t rtt) OVERRIDE;
+  virtual void SetRtt(int64_t rtt) OVERRIDE;
   virtual RateControlRegion Update(const RateControlInput* input,
                                    int64_t now_ms) OVERRIDE;
   virtual void SetEstimate(int bitrate_bps, int64_t now_ms) OVERRIDE;
@@ -58,7 +58,7 @@ class AimdRateControl : public RemoteRateControl {
   uint32_t MultiplicativeRateIncrease(int64_t now_ms, int64_t last_ms,
                                       uint32_t current_bitrate_bps) const;
   uint32_t AdditiveRateIncrease(int64_t now_ms, int64_t last_ms,
-                                uint32_t response_time_ms) const;
+                                int64_t response_time_ms) const;
   void UpdateChangePeriod(int64_t now_ms);
   void UpdateMaxBitRateEstimate(float incoming_bit_rate_kbps);
   void ChangeState(const RateControlInput& input, int64_t now_ms);
@@ -80,7 +80,7 @@ class AimdRateControl : public RemoteRateControl {
   int64_t time_first_incoming_estimate_;
   bool bitrate_is_initialized_;
   float beta_;
-  uint32_t rtt_;
+  int64_t rtt_;
   int64_t time_of_last_log_;
 
   DISALLOW_IMPLICIT_CONSTRUCTORS(AimdRateControl);
